my-turborepo/apps/py-api/pyproject.toml
[project]
name = "py-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "pandas>=2.3.3",
    "pyarrow>=22.0.0",
    "taskiq>=0.12.1",
    "uvicorn>=0.40.0",
]
